+++ src/whatslab/solvers/__init__.py
@@ -1,5 +1,6 @@
 from .arm import ArmIK, DiffArmIK, backend_cls, xyzrpy_to_mat
-from .hand import CONFIG_REGISTRY, HandRetargetController, HandRetargeter
+from .hand import CONFIG_REGISTRY, HandRetargetController, HandRetargeter, KPHandRetargeter
 
 __all__ = ["ArmIK", "DiffArmIK", "backend_cls", "xyzrpy_to_mat",
-           "HandRetargeter", "HandRetargetController", "CONFIG_REGISTRY"]
+           "HandRetargeter", "KPHandRetargeter", "HandRetargetController",
+           "CONFIG_REGISTRY"]